Config file now has debug setting.
page_out in jsonSupport has been renamed to ajaxable_page and it
relies on the old page_out which is now in the send mixin

Now possible to keep applications separately from server.  To test
this  uploadPic expander and uploads mixin have been moved to a
test_apps subdirectory.  Also to complete the test css and js
files have been added for uploadPic's use.